project db cleanup + getAllProjectsByUserId

+++ back-end/domain/data-access/project.db.ts
@@ -1,18 +1,12 @@
 import { Project } from "../model/project";
 import prisma from "../../util/init-db";
 
-
 const getAllProjects = async (): Promise<Project[]> => {
     try {
         const projectsPrisma = await prisma.project.findMany(
             {
                 include: {
-                    tasks: true,
-                    team: {
-                        include: {
-                            users: true
-                        }
-                    }
+                    team: true
                 }
             }
         );
@@ -26,18 +20,13 @@ const getAllProjects = async (): Promise<Project[]> => {
         console.log(error);
         throw new Error("Database error. Check logs for more details.");
     }
-    }
+}
 const getProjectById = async (id: number): Promise<Project> => {
     try {
         const projectPrisma = await prisma.project.findUnique({
             where: { id: id },
             include: {
-                team: {
-                    include: {
-                        users: true
-                    }
-                },
-                tasks: true
+                team: true
             }
         });
         if (!projectPrisma) throw new Error(`Project with id ${id} does not exist.`);
@@ -63,16 +52,10 @@ const createProject = async (project: Project): Promise<Project> => {
                 }
             };
         }
-        // console.log(data);
         const newProject = await prisma.project.create({
             data,
             include: {
-                team: {
-                    include: {
-                        users: true
-                    }
-                },
-                tasks: true
+                team: true
             }
         });
         return Project.from(newProject);
@@ -89,12 +72,7 @@ const deleteById = async (id: number): Promise<Project> => {
         const projectPrisma = await prisma.project.delete({
             where: { id: id },
             include: {
-                tasks: true,
-                team: {
-                    include: {
-                        users: true
-                    }
-                }
+                team: true
             }
         });
         return Project.from(projectPrisma);
@@ -105,5 +83,30 @@ const deleteById = async (id: number): Promise<Project> => {
     }
 }
 
+const getAllProjectsByUserId = async (userId: number): Promise<Project[]> => {
+    try {
+        const projectsPrisma = await prisma.project.findMany({
+            where: {
+                team: {
+                    users: {
+                        some: {
+                            id: userId
+                        }
+                    }
+                }
+            },
+            include: {
+                team: true
+            }
+        });
+        if (!projectsPrisma) throw new Error(`No projects found.`);
+        const projects = projectsPrisma.map((projectPrisma) => Project.from(projectPrisma));
+        return projects;
+    }
+    catch (error) {
+        console.log(error);
+        throw new Error("Database error. Check logs for more details.");
+    }
+}
 
-export default { getAllProjects, getProjectById, createProject, deleteById };
+export default { getAllProjects, getProjectById, createProject, deleteById, getAllProjectsByUserId};
